requirements.txt
zmq
numpy
datetime
pyzmq
scipy
pyserial
matplotlib
opencv-python
